pyproject.toml
[tool.poetry]
name = "ptb_mongo_docker_preset"
version = "0.1.0"
description = ""
authors = ["Anton <109826223+Arcadiyyyyyyyy@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "ptb_mongo_docker_preset"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"
pymongo = "^4.3.3"
python-i18n = "^0.3.9"
flake8 = "3.9.2"
tox = "3.24.3"
pytest = "6.2.5"
pytest-cov = "2.12.1"
mypy = "=0.910"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--cov=ptb_mongo_docker_preset"
testpaths = [
    "tests",
]

[tool.mypy]
mypy_path = "ptb_mongo_docker_preset"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true
